+++ packages/types/src/codec/Tuple.ts
@@ -36,6 +36,7 @@ export default class Tuple extends AbstractArray<Codec> {
     this._Types = Clazzes;
   }
 
+  /** @internal */
   private static decodeTuple (registry: Registry, _Types: TupleConstructors, value: AnyU8a | string | (AnyU8a | AnyNumber | AnyString | undefined | null)[]): Codec[] {
     if (isU8a(value)) {
       return decodeU8a(registry, value, _Types);