+++ packages/types/src/interfaces/runtime/definitions.ts
@@ -2,8 +2,30 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+// Extrinsic definitions, extracted for readability
+const extrinsic = {
+  Extrinsic: 'GenericExtrinsic',
+  ExtrinsicEra: 'GenericExtrinsicEra',
+  ExtrinsicPayload: 'GenericExtrinsicPayload',
+  ExtrinsicV1: 'GenericExtrinsicV1',
+  ExtrinsicPayloadV1: 'GenericExtrinsicPayloadV1',
+  ExtrinsicSignatureV1: 'GenericExtrinsicSignatureV1',
+  ExtrinsicV2: 'GenericExtrinsicV2',
+  ExtrinsicPayloadV2: 'GenericExtrinsicPayloadV2',
+  ExtrinsicSignatureV2: 'GenericExtrinsicSignatureV2',
+  ExtrinsicV3: 'GenericExtrinsicV3',
+  ExtrinsicPayloadV3: 'GenericExtrinsicPayloadV3',
+  ExtrinsicSignatureV3: 'GenericExtrinsicSignatureV3',
+  ExtrinsicV4: 'GenericExtrinsicV4',
+  ExtrinsicPayloadV4: 'GenericExtrinsicPayloadV4',
+  ExtrinsicSignatureV4: 'GenericExtrinsicSignatureV4',
+  ExtrinsicUnknown: 'GenericExtrinsicUnknown',
+  ExtrinsicPayloadUnknown: 'GenericExtrinsicPayloadUnknown'
+};
+
 export default {
   types: {
+    ...extrinsic,
     AccountId: 'GenericAccountId',
     AccountIdOf: 'AccountId',
     AccountIndex: 'GenericAccountIndex',
@@ -33,19 +55,6 @@ export default {
       weight: 'Weight',
       class: 'DispatchClass'
     },
-    Extrinsic: 'GenericExtrinsic',
-    ExtrinsicEra: 'GenericExtrinsicEra',
-    ExtrinsicPayload: 'GenericExtrinsicPayload',
-    ExtrinsicPayloadUnknown: 'GenericExtrinsicPayloadUnknown',
-    ExtrinsicPayloadV1: 'GenericExtrinsicPayloadV1',
-    ExtrinsicPayloadV2: 'GenericExtrinsicPayloadV2',
-    ExtrinsicPayloadV3: 'GenericExtrinsicPayloadV3',
-    ExtrinsicPayloadV4: 'GenericExtrinsicPayloadV4',
-    ExtrinsicUnknown: 'GenericExtrinsicUnknown',
-    ExtrinsicV1: 'GenericExtrinsicV1',
-    ExtrinsicV2: 'GenericExtrinsicV2',
-    ExtrinsicV3: 'GenericExtrinsicV3',
-    ExtrinsicV4: 'GenericExtrinsicV4',
     Hash: 'H256',
     Header: {
       parentHash: 'Hash',