+++ packages/types/src/primitive/Extrinsic/v4/ExtrinsicPayload.ts
@@ -10,10 +10,9 @@ import Struct from '../../../codec/Struct';
 import Bytes from '../../../primitive/Bytes';
 import u32 from '../../../primitive/U32';
 import { sign } from '../util';
-import { SignedPayloadBaseV2 as SignedPayloadBaseV4 } from '../v2/ExtrinsicPayload';
 
 /**
- * @name ExtrinsicPayloadV4
+ * @name GenericExtrinsicPayloadV4
  * @description
  * A signing payload for an [[Extrinsic]]. For the final encoding, it is variable length based
  * on the contents included
@@ -21,19 +20,9 @@ import { SignedPayloadBaseV2 as SignedPayloadBaseV4 } from '../v2/ExtrinsicPaylo
 export default class ExtrinsicPayloadV4 extends Struct {
   constructor (registry: Registry, value?: ExtrinsicPayloadValue | Uint8Array | string) {
     super(registry, {
-      ...SignedPayloadBaseV4,
-      // TODO get the signed extensions from the registry and apply
-      ...{
-        // system::CheckVersion<Runtime>
-        specVersion: 'u32',
-        // system::CheckGenesis<Runtime>
-        genesisHash: 'Hash',
-        // system::CheckEra<Runtime>
-        blockHash: 'Hash'
-        // system::CheckNonce<Runtime>
-        // system::CheckWeight<Runtime>
-        // balances::TakeFees<Runtime>
-      }
+      method: 'Bytes',
+      ...registry.getSignedExtensionTypes(),
+      ...registry.getSignedExtensionExtra()
     }, value);
   }