+++ packages/types/src/types.ts
@@ -372,6 +372,12 @@ export interface RegistryMetadataEvents {
   unwrap (): RegistryMetadataEvent[];
 }
 
+export interface RegistryMetadataExtrinsic {
+  version: BN;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  signedExtensions: String[];
+}
+
 export interface RegistryMetadataModule {
   calls: RegistryMetadataCalls;
   errors: RegistryMetadataErrors;
@@ -381,6 +387,7 @@ export interface RegistryMetadataModule {
 
 export interface RegistryMetadataLatest {
   modules: RegistryMetadataModule[];
+  extrinsic: RegistryMetadataExtrinsic;
 }
 
 export interface RegistryMetadata {
@@ -396,6 +403,8 @@ export interface Registry {
 
   get <T extends Codec = Codec> (name: string): Constructor<T> | undefined;
   getOrThrow <T extends Codec = Codec> (name: string, msg?: string): Constructor<T>;
+  getSignedExtensionExtra (): Record<string, InterfaceTypes>;
+  getSignedExtensionTypes (): Record<string, InterfaceTypes>;
   hasClass (name: string): boolean;
   hasDef (name: string): boolean;
   hasType (name: string): boolean;